+++ app/src/main/java/openfoodfacts/github/scrachx/openfood/features/PreferencesFragment.kt
@@ -17,6 +17,7 @@ package openfoodfacts.github.scrachx.openfood.features
 
 import android.content.ActivityNotFoundException
 import android.content.Intent
+import android.content.Intent.ACTION_VIEW
 import android.content.SharedPreferences
 import android.content.SharedPreferences.OnSharedPreferenceChangeListener
 import android.content.pm.PackageManager
@@ -28,6 +29,7 @@ import android.util.Log
 import android.view.Menu
 import android.view.MenuInflater
 import android.widget.Toast
+import androidx.annotation.StringRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.browser.customtabs.CustomTabsIntent
@@ -61,20 +63,22 @@ import openfoodfacts.github.scrachx.openfood.models.entities.analysistagconfig.A
 import openfoodfacts.github.scrachx.openfood.models.entities.analysistagconfig.AnalysisTagConfigDao
 import openfoodfacts.github.scrachx.openfood.models.entities.country.CountryName
 import openfoodfacts.github.scrachx.openfood.models.entities.country.CountryNameDao
-import openfoodfacts.github.scrachx.openfood.utils.*
+import openfoodfacts.github.scrachx.openfood.utils.INavigationItem
 import openfoodfacts.github.scrachx.openfood.utils.LocaleHelper.getLanguage
 import openfoodfacts.github.scrachx.openfood.utils.LocaleHelper.getLocale
+import openfoodfacts.github.scrachx.openfood.utils.NavigationDrawerListener
 import openfoodfacts.github.scrachx.openfood.utils.NavigationDrawerListener.NavigationDrawerType
+import openfoodfacts.github.scrachx.openfood.utils.SearchSuggestionProvider
+import openfoodfacts.github.scrachx.openfood.utils.requirePreference
 import org.greenrobot.greendao.async.AsyncOperation
 import org.greenrobot.greendao.async.AsyncOperationListener
 import org.greenrobot.greendao.query.WhereCondition.StringCondition
 import java.util.*
 
-/**
- * A class for creating all the ListPreference
- */
+
 class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnSharedPreferenceChangeListener {
     private val disp = CompositeDisposable()
+
     override val navigationDrawerListener: NavigationDrawerListener? by lazy {
         if (activity is NavigationDrawerListener) activity as NavigationDrawerListener
         else null
@@ -90,54 +94,54 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
 
         val settings = requireActivity().getSharedPreferences("prefs", 0)
 
-        val finalLocalValues = arrayListOf<String>()
-        val finalLocalLabels = arrayListOf<String?>()
+        val finalLocalLangs = mutableListOf<String>()
+        val finalLocalLabels = mutableListOf<String?>()
 
-        val localeValues = requireActivity().resources.getStringArray(R.array.languages_array)
-        val localeLabels = arrayOfNulls<String>(localeValues.size)
+        val languages = requireActivity().resources.getStringArray(R.array.languages_array)
+        val localeLabels = arrayOfNulls<String>(languages.size)
 
-        localeValues.withIndex().forEach { (i, value) ->
-            val current = getLocale(value)
+        languages.withIndex().forEach { (i, lang) ->
+            val current = getLocale(lang)
             localeLabels[i] = current.getDisplayName(current).capitalize(Locale.getDefault())
-            finalLocalLabels.add(localeLabels[i])
-            finalLocalValues.add(value)
-
-        }
-        val languagePreference = requirePreference<ListPreference>("Locale.Helper.Selected.Language")
-        languagePreference.entries = finalLocalLabels.toTypedArray()
-        languagePreference.entryValues = finalLocalValues.toTypedArray()
-        languagePreference.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _: Preference?, locale: Any? ->
-            val configuration = requireActivity().resources.configuration
-            Toast.makeText(context, getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
-                configuration.setLocale(getLocale(locale as String?))
-                requireActivity().recreate()
-            }
-            true
+            finalLocalLabels += localeLabels[i]
+            finalLocalLangs += lang
         }
 
-        val appThemeEntries = resources.getStringArray(R.array.application_theme_entries)
-        requirePreference<ListPreference>("applicationThemePreference").let {
-            it.setEntries(R.array.application_theme_entries)
-            it.setEntryValues(R.array.application_theme_entries)
-            it.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, value ->
-                when (value) {
-                    appThemeEntries[1] -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-                    appThemeEntries[2] -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-                    else -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM)
+        requirePreference<ListPreference>(getString(R.string.pref_language_key)).let {
+            it.entries = finalLocalLabels.toTypedArray()
+            it.entryValues = finalLocalLangs.toTypedArray()
+            it.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _: Preference?, locale: Any? ->
+                val configuration = requireActivity().resources.configuration
+                Toast.makeText(context, getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN_MR1) {
+                    configuration.setLocale(getLocale(locale as String?))
+                    requireActivity().recreate()
                 }
                 true
             }
         }
 
-        requirePreference<Preference>("deleteSearchHistoryPreference").onPreferenceClickListener = OnPreferenceClickListener {
+
+        requirePreference<ListPreference>(getString(R.string.pref_app_theme_key)).let {
+            it.setEntries(R.array.application_theme_entries)
+            it.setEntryValues(R.array.application_theme_entries)
+            it.setOnPreferenceChangeListener { _, value ->
+                when (value) {
+                    getString(R.string.day) -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
+                    getString(R.string.night) -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
+                    getString(R.string.follow_system) -> AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM)
+                }
+                true
+            }
+        }
+
+        requirePreference<Preference>(getString(R.string.pref_delete_history_key)).setOnPreferenceChangeListener { _, _ ->
             MaterialDialog.Builder(requireActivity()).run {
-                content(R.string.search_history_pref_dialog_content)
+                content(R.string.pref_delete_history_dialog_content)
                 positiveText(R.string.delete_txt)
                 onPositive { _, _ ->
-                    Toast.makeText(context, getString(R.string.preference_delete_search_history), Toast.LENGTH_SHORT).show()
-                    val suggestions = SearchRecentSuggestions(context, SearchSuggestionProvider.AUTHORITY, SearchSuggestionProvider.MODE)
-                    suggestions.clearHistory()
+                    Toast.makeText(requireContext(), getString(R.string.pref_delete_history), Toast.LENGTH_SHORT).show()
+                    SearchRecentSuggestions(requireContext(), SearchSuggestionProvider.AUTHORITY, SearchSuggestionProvider.MODE).clearHistory()
                 }
                 neutralText(R.string.dialog_cancel)
                 onNeutral { dialog, _ -> dialog.dismiss() }
@@ -149,7 +153,7 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
         val countryLabels = mutableListOf<String>()
         val countryTags = mutableListOf<String>()
 
-        val countryPreference = requirePreference<ListPreference>(LocaleHelper.USER_COUNTRY_PREFERENCE_KEY)
+        val countryPreference = requirePreference<ListPreference>(getString(R.string.pref_country_key))
 
         val asyncSessionCountries = OFFApplication.daoSession.startAsyncSession()
         val countryNameDao = OFFApplication.daoSession.countryNameDao
@@ -157,9 +161,9 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
         // Set query finish listener
         asyncSessionCountries.listenerMainThread = AsyncOperationListener { operation: AsyncOperation ->
             val countryNames = operation.result as List<CountryName>
-            countryNames.forEach { countryName ->
-                countryLabels.add(countryName.name)
-                countryTags.add(countryName.countyTag)
+            countryNames.forEach {
+                countryLabels.add(it.name)
+                countryTags.add(it.countyTag)
             }
             countryPreference.entries = countryLabels.toTypedArray()
             countryPreference.entryValues = countryTags.toTypedArray()
@@ -169,15 +173,14 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
                 .where(CountryNameDao.Properties.LanguageCode.eq(getLanguage(requireActivity())))
                 .orderAsc(CountryNameDao.Properties.Name).build())
 
-        countryPreference.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { preference: Preference, newValue: Any? ->
-            if (preference is ListPreference && preference.getKey() == LocaleHelper.USER_COUNTRY_PREFERENCE_KEY) {
-                val country = newValue as String?
-                settings.edit { putString(preference.getKey(), country) }
-                Toast.makeText(context, getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
-            }
+        countryPreference.setOnPreferenceChangeListener { preference, newValue ->
+            val country = newValue as String?
+            settings.edit { putString(preference.key, country) }
+            Toast.makeText(context, getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
             true
         }
-        requirePreference<Preference>("contact_team").onPreferenceClickListener = OnPreferenceClickListener {
+
+        requirePreference<Preference>(getString(R.string.pref_contact_us_key)).setOnPreferenceChangeListener { _, _ ->
             val contactIntent = Intent(Intent.ACTION_SENDTO)
             contactIntent.data = Uri.parse(getString(R.string.off_mail))
             contactIntent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
@@ -188,44 +191,49 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
             }
             true
         }
-        requirePreference<Preference>("RateUs").onPreferenceClickListener = OnPreferenceClickListener {
+        requirePreference<Preference>(getString(R.string.pref_rate_us_key)).setOnPreferenceChangeListener { _, _ ->
             try {
-                startActivity(Intent(Intent.ACTION_VIEW, Uri.parse("market://details?id=" + requireActivity().packageName)))
+                startActivity(Intent(ACTION_VIEW, Uri.parse("market://details?id=${requireActivity().packageName}")))
             } catch (e: ActivityNotFoundException) {
-                startActivity(Intent(Intent.ACTION_VIEW, Uri.parse("https://play.google.com/store/apps/details?id=" + requireActivity().packageName)))
+                startActivity(Intent(ACTION_VIEW, Uri.parse("https://play.google.com/store/apps/details?id=${requireActivity().packageName}")))
             }
             true
         }
-        requirePreference<Preference>("FAQ").onPreferenceClickListener =
-                OnPreferenceClickListener { openWebCustomTab(R.string.faq_url) }
-        requirePreference<Preference>("Terms").onPreferenceClickListener =
-                OnPreferenceClickListener { openWebCustomTab(R.string.terms_url) }
-        requirePreference<Preference>("local_translate_help").onPreferenceClickListener =
-                OnPreferenceClickListener { openWebCustomTab(R.string.translate_url) }
-        val energyUnitPreference = requirePreference<ListPreference>("energyUnitPreference")
-        val energyUnits = requireActivity().resources.getStringArray(R.array.energy_units)
-        energyUnitPreference.entries = energyUnits
-        energyUnitPreference.entryValues = energyUnits
-        energyUnitPreference.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
-            settings.edit { putString("energyUnitPreference", newValue as String?) }
-            Toast.makeText(requireActivity(), getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
-            true
-        }
-        val volumeUnitPreference = requirePreference<ListPreference>("volumeUnitPreference")
-        val volumeUnits = requireActivity().resources.getStringArray(R.array.volume_units)
-        volumeUnitPreference.entries = volumeUnits
-        volumeUnitPreference.entryValues = volumeUnits
-        volumeUnitPreference.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
-            settings.edit { putString("volumeUnitPreference", newValue as String?) }
-            Toast.makeText(requireActivity(), getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
-            true
-        }
-        val values = requireActivity().resources.getStringArray(R.array.upload_image)
-        requirePreference<ListPreference>("ImageUpload").let {
-            it.entries = values
-            it.entryValues = values
+
+        requirePreference<Preference>(getString(R.string.pref_faq_key))
+                .setOnPreferenceClickListener { openWebCustomTab(R.string.faq_url) }
+
+        requirePreference<Preference>(getString(R.string.pref_terms_key))
+                .setOnPreferenceClickListener { openWebCustomTab(R.string.terms_url) }
+
+        requirePreference<Preference>(getString(R.string.pref_help_translate_key))
+                .setOnPreferenceClickListener { openWebCustomTab(R.string.translate_url) }
+
+        requirePreference<ListPreference>(getString(R.string.pref_energy_unit_key)).let {
+            it.setEntries(R.array.energy_units)
+            it.setEntryValues(R.array.energy_units)
             it.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
-                settings.edit { putString("imageUpload", newValue as String?) }
+                settings.edit { putString(getString(R.string.pref_energy_unit_key), newValue as String?) }
+                Toast.makeText(requireActivity(), getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
+                true
+            }
+        }
+
+        requirePreference<ListPreference>(getString(R.string.pref_volume_unit_key)).let {
+            it.setEntries(R.array.volume_units)
+            it.setEntryValues(R.array.volume_units)
+            it.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
+                settings.edit { putString(getString(R.string.pref_volume_unit_key), newValue as String?) }
+                Toast.makeText(requireActivity(), getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
+                true
+            }
+        }
+
+        requirePreference<ListPreference>(getString(R.string.pref_resolution_key)).let {
+            it.setEntries(R.array.upload_image)
+            it.setEntryValues(R.array.upload_image)
+            it.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, newValue ->
+                settings.edit { putString(getString(R.string.pref_resolution_key), newValue as String?) }
                 Toast.makeText(requireActivity(), getString(R.string.changes_saved), Toast.LENGTH_SHORT).show()
                 true
             }
@@ -233,11 +241,10 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
 
 
         // Disable photo mode for OpenProductFacts
-        if (isFlavors(AppFlavors.OPF)) requirePreference<Preference>("photoMode").isVisible = false
+        if (isFlavors(AppFlavors.OPF)) requirePreference<Preference>(getString(R.string.pref_show_product_photos_key)).isVisible = false
 
         // Preference to show version name
-        requirePreference<Preference>("Version").let {
-            it.isEnabled = false
+        requirePreference<Preference>(getString(R.string.pref_version_key)).let {
             try {
                 val pInfo = requireActivity().packageManager.getPackageInfo(requireActivity().packageName, 0)
                 val version = pInfo.versionName
@@ -246,10 +253,11 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
             } catch (e: PackageManager.NameNotFoundException) {
                 Log.e(PreferencesFragment::class.simpleName, "onCreatePreferences", e)
             }
+
             if (isFlavors(OFF, OBF, OPFF)) {
                 getAnalysisTagConfigs(OFFApplication.daoSession)
             } else {
-                preferenceScreen.removePreference(preferenceScreen.requirePreference("display_category"))
+                preferenceScreen.removePreference(preferenceScreen.requirePreference(getString(R.string.pref_key_display)))
             }
         }
     }
@@ -257,8 +265,7 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
     private fun buildDisplayCategory(configs: List<AnalysisTagConfig>?) {
         if (!isAdded) return
 
-        val displayCategory = preferenceScreen.findPreference<PreferenceCategory>("display_category")
-                ?: throw error("Display category preference does not exist.")
+        val displayCategory = preferenceScreen.requirePreference<PreferenceCategory>(getString(R.string.pref_key_display))
         displayCategory.removeAll()
         preferenceScreen.addPreference(displayCategory)
 
@@ -274,7 +281,7 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
 
                 // The service will load server resources only if newer than already downloaded...
                 manager.enqueue(request)
-                manager.getWorkInfoByIdLiveData(request.id).observe(this@PreferencesFragment, { workInfo: WorkInfo? ->
+                manager.getWorkInfoByIdLiveData(request.id).observe(this, { workInfo: WorkInfo? ->
                     if (workInfo != null) {
                         if (workInfo.state == WorkInfo.State.RUNNING) {
                             preference.setTitle(R.string.please_wait)
@@ -291,7 +298,7 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
             displayCategory.addPreference(preference)
         } else {
             configs.forEach { config ->
-                displayCategory.addPreference(CheckBoxPreference(this.preferenceScreen.context).apply {
+                displayCategory.addPreference(CheckBoxPreference(this.context).apply {
                     key = config.type
                     setDefaultValue(true)
                     summary = null
@@ -304,10 +311,10 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
         displayCategory.isVisible = true
     }
 
-    private fun openWebCustomTab(faqUrl: Int): Boolean {
+    private fun openWebCustomTab(@StringRes resId: Int): Boolean {
         val customTabsIntent = CustomTabsIntent.Builder().build()
-        customTabsIntent.intent.putExtra("android.intent.extra.REFERRER", Uri.parse("android-app://" + requireContext().packageName))
-        CustomTabActivityHelper.openCustomTab(requireActivity(), customTabsIntent, Uri.parse(getString(faqUrl)), WebViewFallback())
+        customTabsIntent.intent.putExtra("android.intent.extra.REFERRER", Uri.parse("android-app://${requireContext().packageName}"))
+        CustomTabActivityHelper.openCustomTab(requireActivity(), customTabsIntent, Uri.parse(getString(resId)), WebViewFallback())
         return true
     }
 
@@ -319,7 +326,7 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
         try {
             (this.activity as? AppCompatActivity)?.supportActionBar!!.title = getString(R.string.action_preferences)
         } catch (e: NullPointerException) {
-            Log.e(this::class.simpleName, "on resume error", e)
+            throw IllegalStateException("Preference fragment not attached to AppCompatActivity.")
         }
         preferenceManager.sharedPreferences.registerOnSharedPreferenceChangeListener(this)
     }
@@ -336,7 +343,7 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
 
     override fun onSharedPreferenceChanged(sharedPreferences: SharedPreferences, key: String) {
         when (key) {
-            "enableMobileDataUpload" -> scheduleSync()
+            getString(R.string.pref_enable_mobile_data_key) -> scheduleSync()
         }
     }
 
@@ -348,19 +355,17 @@ class PreferencesFragment : PreferenceFragmentCompat(), INavigationItem, OnShare
                     .where(StringCondition("1 GROUP BY type"))
                     .orderAsc(AnalysisTagConfigDao.Properties.Type).build().list()
             val analysisTagNameDao = daoSession.analysisTagNameDao
-            for (config in analysisTagConfigs) {
+            analysisTagConfigs.forEach { config ->
                 val type = "en:${config.type}"
-                var analysisTagTypeName = analysisTagNameDao.queryBuilder()
-                        .where(
-                                AnalysisTagNameDao.Properties.AnalysisTag.eq(type),
-                                AnalysisTagNameDao.Properties.LanguageCode.eq(language),
-                        ).unique()
+                var analysisTagTypeName = analysisTagNameDao.queryBuilder().where(
+                        AnalysisTagNameDao.Properties.AnalysisTag.eq(type),
+                        AnalysisTagNameDao.Properties.LanguageCode.eq(language),
+                ).unique()
                 if (analysisTagTypeName == null) {
-                    analysisTagTypeName = analysisTagNameDao.queryBuilder()
-                            .where(
-                                    AnalysisTagNameDao.Properties.AnalysisTag.eq(type),
-                                    AnalysisTagNameDao.Properties.LanguageCode.eq("en")
-                            ).unique()
+                    analysisTagTypeName = analysisTagNameDao.queryBuilder().where(
+                            AnalysisTagNameDao.Properties.AnalysisTag.eq(type),
+                            AnalysisTagNameDao.Properties.LanguageCode.eq("en")
+                    ).unique()
                 }
                 config.typeName = if (analysisTagTypeName != null) analysisTagTypeName.name else config.type
             }