+++ app/src/main/java/openfoodfacts/github/scrachx/openfood/features/adapters/ProductsRecyclerViewAdapter.kt
@@ -10,6 +10,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.squareup.picasso.Callback
 import com.squareup.picasso.Picasso
+import openfoodfacts.github.scrachx.openfood.AppFlavors
 import openfoodfacts.github.scrachx.openfood.R
 import openfoodfacts.github.scrachx.openfood.models.Product
 import openfoodfacts.github.scrachx.openfood.network.OpenFoodAPIClient.Companion.getLocaleProductNameField
@@ -80,23 +81,22 @@ class ProductsRecyclerViewAdapter(
         // Set product description
         holder.productDetails.text = product?.getProductBrandsQuantityDetails()
 
-        // Set nutriscore icon
-        val nutriRes = product.getNutriScoreResource()
-        holder.productNutriscore.setImageResource(nutriRes)
-        holder.productNutriscore.visibility = View.VISIBLE
+        if (AppFlavors.isFlavors(AppFlavors.OFF)) {
+            // Set nutriscore icon
+            val nutriRes = product.getNutriScoreResource()
+            holder.productNutriscore.setImageResource(nutriRes)
+            holder.productNutriscore.visibility = View.VISIBLE
 
+            // Set nova icon
+            val novaRes = product.getNovaGroupResource()
+            holder.productNova.setImageResource(novaRes)
+            holder.productNova.visibility = View.VISIBLE
 
-        // Set nova icon
-        val novaRes = product.getNovaGroupResource()
-        holder.productNova.setImageResource(novaRes)
-        holder.productNova.visibility = View.VISIBLE
-
-
-        // Set ecoscore icon
-        val ecoRes = product.getEcoscoreResource()
-        holder.productEcoscore.setImageResource(ecoRes)
-        holder.productEcoscore.visibility = View.VISIBLE
-
+            // Set ecoscore icon
+            val ecoRes = product.getEcoscoreResource()
+            holder.productEcoscore.setImageResource(ecoRes)
+            holder.productEcoscore.visibility = View.VISIBLE
+        }
     }
 
     fun getProduct(position: Int) = products[position]