+++ app/src/main/java/openfoodfacts/github/scrachx/openfood/features/product/view/summary/SummaryProductFragment.kt
@@ -53,6 +53,7 @@ import openfoodfacts.github.scrachx.openfood.features.LoginActivity.Companion.Lo
 import openfoodfacts.github.scrachx.openfood.features.additives.AdditiveFragmentHelper.showAdditives
 import openfoodfacts.github.scrachx.openfood.features.compare.ProductCompareActivity.Companion.start
 import openfoodfacts.github.scrachx.openfood.features.product.edit.ProductEditActivity
+import openfoodfacts.github.scrachx.openfood.features.product.edit.ProductEditActivity.Companion.KEY_STATE
 import openfoodfacts.github.scrachx.openfood.features.product.view.CategoryProductHelper
 import openfoodfacts.github.scrachx.openfood.features.product.view.ProductViewActivity
 import openfoodfacts.github.scrachx.openfood.features.product.view.ingredients_analysis.IngredientsWithTagDialogFragment
@@ -125,6 +126,14 @@ class SummaryProductFragment : BaseFragment(), ISummaryProductPresenter.View {
     /**boolean to determine if nutrient prompt should be shown*/
     private var showNutrientPrompt = false
 
+    /**boolean to determine if eco score prompt should be shown*/
+    private var showEcoScorePrompt = false
+
+    /**boolean to determine if labels prompt should be shown*/
+    private var showLabelsPrompt = false
+
+    /**boolean to determine if origins prompt should be shown*/
+    private var showOriginsPrompt = false
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -247,7 +256,7 @@ class SummaryProductFragment : BaseFragment(), ISummaryProductPresenter.View {
         binding.labelsIcon.visibility = View.VISIBLE
 
         // Checks the product states_tags to determine which prompt to be shown
-        refreshNutriScorePrompt()
+        refreshStatesTagsPrompt()
         presenter.loadAllergens(null)
         presenter.loadCategories()
         presenter.loadLabels()
@@ -446,21 +455,36 @@ class SummaryProductFragment : BaseFragment(), ISummaryProductPresenter.View {
 
     private fun refreshCO2OrEcoscoreIcon() {
         binding.ecoscoreIcon.setImageResource(product.getEcoscoreResource())
+        binding.ecoscoreIcon.setOnClickListener{
+            val uri = Uri.parse(getString(R.string.ecoscore_url))
+            val customTabsIntent = CustomTabsHelper.getCustomTabsIntent(requireContext(), customTabActivityHelper.session)
+            CustomTabActivityHelper.openCustomTab(requireActivity(), customTabsIntent, uri, WebViewFallback())
+        }
     }
 
-    private fun refreshNutriScorePrompt() {
+    private fun refreshStatesTagsPrompt() {
         //checks the product states_tags to determine which prompt to be shown
         val statesTags = product.statesTags
         showCategoryPrompt = statesTags.contains("en:categories-to-be-completed") && !hasCategoryInsightQuestion
         showNutrientPrompt = statesTags.contains("en:nutrition-facts-to-be-completed") && product.noNutritionData != "on"
+        showEcoScorePrompt = statesTags.contains("en:categories-completed") && (product.ecoscore.isNullOrEmpty() || product.ecoscore.equals("unknown", true))
+        showLabelsPrompt = statesTags.contains("en:labels-to-be-completed")
+        showOriginsPrompt = statesTags.contains("en:origins-to-be-completed")
 
         Log.d(LOG_TAG, "Show category prompt: $showCategoryPrompt")
         Log.d(LOG_TAG, "Show nutrient prompt: $showNutrientPrompt")
+        Log.d(LOG_TAG, "Show Eco Score prompt: $showEcoScorePrompt")
+        Log.d(LOG_TAG, "Show labels prompt: $showLabelsPrompt")
+        Log.d(LOG_TAG, "Show origins prompt: $showOriginsPrompt")
+
+        if(showEcoScorePrompt){
+            binding.tipBoxEcoScore.loadToolTip()
+        }
 
         binding.addNutriscorePrompt.visibility = View.VISIBLE
         when {
             showNutrientPrompt && showCategoryPrompt -> {
-                // Both true
+                // showNutrientPrompt and showCategoryPrompt true
                 binding.addNutriscorePrompt.text = getString(R.string.add_nutrient_category_prompt_text)
             }
             showNutrientPrompt -> {
@@ -472,6 +496,23 @@ class SummaryProductFragment : BaseFragment(), ISummaryProductPresenter.View {
                 binding.addNutriscorePrompt.text = getString(R.string.add_category_prompt_text)
             }
             else -> binding.addNutriscorePrompt.visibility = View.GONE
+        }
+
+        binding.addLabelOriginPrompt.visibility = View.VISIBLE
+        when {
+            showLabelsPrompt && showOriginsPrompt -> {
+                // showLabelsPrompt and showOriginsPrompt true
+                binding.addLabelOriginPrompt.text = getString(R.string.add_labels_origins_prompt_text)
+            }
+            showLabelsPrompt -> {
+                // showLabelsPrompt true
+                binding.addLabelOriginPrompt.text = getString(R.string.add_labels_prompt_text)
+            }
+            showOriginsPrompt -> {
+                // showOriginsPrompt true
+                binding.addLabelOriginPrompt.text = getString(R.string.add_origins_prompt_text)
+            }
+            else -> binding.addLabelOriginPrompt.visibility = View.GONE
         }
     }
 
@@ -545,7 +586,7 @@ class SummaryProductFragment : BaseFragment(), ISummaryProductPresenter.View {
         }
 
         if (isFlavors(OFF)) {
-            refreshNutriScorePrompt()
+            refreshStatesTagsPrompt()
             refreshScoresLayout()
         }
     }
@@ -835,5 +876,11 @@ class SummaryProductFragment : BaseFragment(), ISummaryProductPresenter.View {
         private const val EDIT_PRODUCT_NUTRITION_AFTER_LOGIN = 3
         private const val EDIT_REQUEST_CODE = 2
         private val LOG_TAG = this::class.simpleName!!
+
+        fun newInstance(productState: ProductState) = SummaryProductFragment().apply {
+            arguments = Bundle().apply {
+                putSerializable(KEY_STATE, productState)
+            }
+        }
     }
 }