+++ app/src/main/java/openfoodfacts/github/scrachx/openfood/features/shared/views/TipBox.kt
@@ -1,12 +1,16 @@
 package openfoodfacts.github.scrachx.openfood.features.shared.views
 
+import android.content.ActivityNotFoundException
 import android.content.Context
+import android.content.Intent
 import android.content.SharedPreferences
+import android.net.Uri
 import android.util.AttributeSet
 import android.view.*
 import android.view.animation.Animation
 import android.view.animation.Transformation
 import android.widget.LinearLayout
+import android.widget.Toast
 import androidx.core.content.edit
 import androidx.core.content.res.ResourcesCompat
 import androidx.core.content.res.getStringOrThrow
@@ -19,7 +23,7 @@ class TipBox(context: Context, attrs: AttributeSet?) : LinearLayout(context, att
 
     private val prefs: SharedPreferences by lazy { PreferenceManager.getDefaultSharedPreferences(getContext()) }
 
-    private val binding = TipBoxBinding.inflate(LayoutInflater.from(context), this, false)
+    private val binding = TipBoxBinding.inflate(LayoutInflater.from(context), this, true)
     private val shouldAnimate: Boolean
     private val identifier: String
 
@@ -40,11 +44,22 @@ class TipBox(context: Context, attrs: AttributeSet?) : LinearLayout(context, att
                 ResourcesCompat.getColor(resources, R.color.brand_light_blue, context.theme))
         attributes.recycle()
 
-        binding.tipMessage.setTextColor(toolTipTextColor)
-        binding.tipMessage.text = context.getString(R.string.tip_message, message)
+        // if identifier != "showEcoScorePrompt" , set values to button Tip Box
+        // else. set value to icon tip box
+        if(identifier!= "showEcoScorePrompt"){
+            binding.tipBoxIcon.visibility= View.GONE
+            binding.tipMessageButton.setTextColor(toolTipTextColor)
+            binding.tipMessageButton.text = context.getString(R.string.tip_message, message)
+            binding.tipBoxContainer.setBackgroundColor(toolTipBackgroundColor)
 
-        binding.arrow.setColorFilter(toolTipBackgroundColor)
-        setArrowAlignment(arrowAlignment, marginStart, marginEnd)
+            binding.arrow.setColorFilter(toolTipBackgroundColor)
+            setArrowAlignment(arrowAlignment, marginStart, marginEnd)
+        } else{
+            binding.tipBoxButton.visibility = View.GONE
+            binding.tipMessageIcon.setTextColor(toolTipTextColor)
+            binding.tipMessageIcon.text = message
+            binding.tipBoxIcon.setBackgroundColor(toolTipBackgroundColor)
+        }
 
         visibility = GONE
 
@@ -53,14 +68,32 @@ class TipBox(context: Context, attrs: AttributeSet?) : LinearLayout(context, att
             prefs.edit { putBoolean(identifier, false) }
         }
 
-        binding.tipBoxContainer.setBackgroundColor(toolTipBackgroundColor)
+        binding.closeBoxIcon.setOnClickListener {
+            hide()
+            prefs.edit { putBoolean(identifier, false) }
+        }
+
+        binding.mail.setOnClickListener {
+            val contactIntent = Intent(Intent.ACTION_SENDTO)
+            contactIntent.data = Uri.parse("mailto:contact@openfoodfacts.org")
+            contactIntent.flags = Intent.FLAG_ACTIVITY_NEW_TASK
+            try {
+                context.startActivity(contactIntent)
+            } catch (e: ActivityNotFoundException) {
+                Toast.makeText(getContext(), R.string.email_not_found, Toast.LENGTH_SHORT).show()
+            }
+        }
+
 
         if (canDisplayImmediately) loadToolTip()
     }
 
 
     fun setTipMessage(message: CharSequence?) {
-        binding.tipMessage.text = message
+        if (identifier != "showEcoScorePrompt")
+            binding.tipMessageButton.text = context.getString(R.string.tip_message, message)
+        else
+            binding.tipMessageIcon.text = message
     }
 
     fun setArrowAlignment(arrowAlignment: Int, marginStart: Int, marginEnd: Int) {